Move grid class to a new module
